backend updates

+++ Services/PostService/postService.cs
@@ -1,8 +1,10 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using Blog.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using postAPI.Models;
 
@@ -13,17 +15,22 @@ namespace Blog.Services.PostService
 
 
         private readonly PostContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public postService(PostContext context)
+        public postService(PostContext context, IHttpContextAccessor httpContextAccessor)
         {
+            _httpContextAccessor = httpContextAccessor;
             _context = context;
 
         }
 
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
         public async Task<ServiceResponse<List<Post>>> AddNewPost(Post newPost)
         {
             ServiceResponse<List<Post>> serviceResponse = new ServiceResponse<List<Post>>();
-            List<Post> dbPost = await _context.Posts.ToListAsync();
+            List<Post> dbPost = await _context.Posts.Where(c => c.User.Id == GetUserId()).ToListAsync();
+            // all users CAN POST A NEW POST
+            newPost.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
             await _context.Posts.AddAsync(newPost);
             await _context.SaveChangesAsync();
             serviceResponse.Data = dbPost;
@@ -35,12 +42,19 @@ namespace Blog.Services.PostService
             ServiceResponse<List<Post>> serviceResponse = new ServiceResponse<List<Post>>();
             try
             {
-                Post post = await _context.Posts.FirstAsync(c => c.Id == id);
-                _context.Posts.Remove(post);
-
-                List<Post> dbPost = await _context.Posts.ToListAsync();
-                serviceResponse.Data = dbPost;
-                await _context.SaveChangesAsync();
+                Post post = await _context.Posts.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if (post != null)
+                {
+                    _context.Posts.Remove(post);
+                    await _context.SaveChangesAsync();
+                    List<Post> dbPost = await _context.Posts.Where(c => c.User.Id == GetUserId()).ToListAsync();
+                    serviceResponse.Data = dbPost;
+                }
+                else
+                {
+                    serviceResponse.Sucsses = false;
+                    serviceResponse.Message = "Post not found!";
+                }
             }
 
             catch (Exception ex)
@@ -57,7 +71,8 @@ namespace Blog.Services.PostService
         public async Task<ServiceResponse<List<Post>>> GetAllPosts()
         {
             ServiceResponse<List<Post>> serviceResponse = new ServiceResponse<List<Post>>();
-            List<Post> dbPost = await _context.Posts.ToListAsync();
+            List<Post> dbPost = await _context.Posts.Where(c => c.User.Id == GetUserId()).ToListAsync();
+            //_context.Posts.Where(c => c.User.Id == userId).ToListAsync(); user can see just his posts
             serviceResponse.Data = dbPost;
             return serviceResponse;
         }
@@ -65,7 +80,7 @@ namespace Blog.Services.PostService
         public async Task<ServiceResponse<Post>> GetPostById(int id)
         {
             ServiceResponse<Post> serviceResponse = new ServiceResponse<Post>();
-            Post dbPost = await _context.Posts.FirstOrDefaultAsync(c => c.Id == id);
+            Post dbPost = await _context.Posts.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             serviceResponse.Data = dbPost;
             return serviceResponse;
         }
@@ -75,18 +90,27 @@ namespace Blog.Services.PostService
             ServiceResponse<Post> serviceResponse = new ServiceResponse<Post>();
             try
             {
-                Post post = await _context.Posts.FirstOrDefaultAsync(c => c.Id == UpdatedPost.Id);
-                post.PostContent = UpdatedPost.PostContent;
-                post.likes = UpdatedPost.likes;
-                post.pressedLK = UpdatedPost.pressedLK;
-                post.pressedSD = UpdatedPost.pressedSD;
-                post.saved = UpdatedPost.saved;
-                post.userNickname = UpdatedPost.userNickname;
+                Post post = await _context.Posts.Include(c => c.User).FirstOrDefaultAsync(c => c.Id == UpdatedPost.Id);
+                if (post.User.Id == GetUserId())
+                {
+                    post.PostContent = UpdatedPost.PostContent;
+                    post.likes = UpdatedPost.likes;
+                    post.pressedLK = UpdatedPost.pressedLK;
+                    post.pressedSD = UpdatedPost.pressedSD;
+                    post.saved = UpdatedPost.saved;
+                    post.userNickname = UpdatedPost.userNickname;
 
-                _context.Posts.Update(post);
-                await _context.SaveChangesAsync();
+                    _context.Posts.Update(post);
+                    await _context.SaveChangesAsync();
 
-                serviceResponse.Data = post;
+                    serviceResponse.Data = post;
+                }
+                else
+                {
+                    serviceResponse.Sucsses = false;
+                    serviceResponse.Message = "Post not found!";
+                }
+
             }
 
             catch (Exception ex)
